ui: Add Settings + update_rate setting
This allows for adding new user settings in the future.

As a first setting there's update rate, which allows for customizing the
delay between egui repaints.
To be able to pass settings to View UIs there's a new struct, `Viewer`,
which gets created temporarily just to get passed to DockArea::view.
In its TabViewer::ui implementation, it passes the Settings ref to the
Inspector methods.
Alternatively Inspector would have to have a reference to the settings
stored in `App`, but that would result in `App` being self-referential.

Profiler respects the update_rate setting, and no longer refreshes on
user input.
Fixes #3
Fixes #5