backend/pipewire: Drop proxy stores after proxies
The listener callbacks allow for creating reference cycles between
the closures and their owner. When the mainloop quits, the HashMaps
storing the listeners are dropped first and then the PipeWire core.
The core then calls remove on all the proxies causing a use after free
when the last one is dropped:
1. RefCell borrow_mut returns a RefMut guard
2. The proxy and the listener is removed from the store and dropped
3. During that drop, the last Rc to the RefCell containing the map gets
dropped, so it drops them
4. The RefMut from step 1 is dropped, and writes to the RefCell, but it
has been freed

This can be worked around in two ways,
either use Weak in the callback closures
or drop the stores after the proxies

Related https://gitlab.freedesktop.org/pipewire/pipewire-rs/-/issues/109